Indicator rewritten to TS
