+++ src/components/Indicator/Indicator.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import Indicator from '@components/Indicator'
+import { Indicator } from '@components/Indicator'
 import Button from '@components/Button'
 import Icon from '@components/Icon'
