+++ src/components/Indicator/Indicator.test.tsx
@@ -1,0 +1,8 @@
+import { Indicator } from '@components/Indicator'
+import { render } from '@testing-library/react'
+
+describe('<Indicator/> mount', () => {
+  it('is mounted', () => {
+    render(<Indicator>Indicator</Indicator>)
+  })
+})
